requirements.txt
aerich==0.5.3
aiosqlite==0.16.1
asyncpg==0.22.0
click==7.1.2
ddlparse==1.9.0
dictdiffer==0.8.1
fastapi==0.64.0
h11==0.12.0
iso8601==0.1.14
pydantic==1.8.1
pyparsing==2.4.7
pypika-tortoise==0.1.0
pytz==2020.5
starlette==0.13.6
tortoise-orm==0.17.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
